refactor: check element updateGameSettingModalMultilingualContent, feat: components/game/game.js
